refactor: split serializers into different files.
